+++ solvers/ADRSolver/EquationSystems/UnsteadyDiffusion.h
@@ -65,10 +65,14 @@ public:
     ~UnsteadyDiffusion() override;
 
 protected:
+    NekDouble m_epsilon;
+    NekDouble m_d00 = 1.0, m_d11 = 1.0, m_d22 = 1.0;
     bool m_useSpecVanVisc;
-    NekDouble
-        m_sVVCutoffRatio; // cut off ratio from which to start decayhing modes
-    NekDouble m_sVVDiffCoeff; // Diffusion coefficient of SVV modes
+    // cut off ratio from which to start decayhing modes
+    NekDouble m_sVVCutoffRatio;
+    // Diffusion coefficient of SVV modes
+    NekDouble m_sVVDiffCoeff;
+    StdRegions::VarCoeffMap m_varcoeff;
     SolverUtils::DiffusionSharedPtr m_diffusion;
     SolverUtils::RiemannSolverSharedPtr m_riemannSolver;
 
@@ -93,12 +97,6 @@ protected:
         const Array<OneD, const Array<OneD, NekDouble>> &inarray,
         Array<OneD, Array<OneD, NekDouble>> &outarray, NekDouble time,
         NekDouble lambda);
-
-private:
-    NekDouble m_waveFreq;
-    NekDouble m_epsilon;
-    NekDouble m_d00 = 1.0, m_d11 = 1.0, m_d22 = 1.0;
-    StdRegions::VarCoeffMap m_varcoeff;
 };
 } // namespace Nektar